+++ domains/certificates/Query.API/API.AppTests/SwaggerTests.GetSwaggerDoc_AppStarted_NoChangesAccordingToSnapshot.verified.txt
@@ -222,7 +222,7 @@
     "/api/certificates/owner/uuid": {
       "get": {
         "tags": [
-          "TransferCertificate"
+          "Uuid"
         ],
         "responses": {
           "200": {
